Add files via upload
Tehran is the capital of Iran and the largest and the most populous city in Iran. The city's daily population is estimated at 14 million. There are 22 million trips per day in the city. This city has been selected in many research studies because of its political, economic, and geopolitical significance. To help researchers, we provided a dataset including all geographic information of intersections, links and their lengths along with an intermediate point on each link. This network consists of 13743 links and 7784 nodes.